[Feature][USER_005.01] List waiting interview

+++ config/validate.php
@@ -12,5 +12,5 @@ return [
     'phone_min_length' => 10,
     'phone_max_length' => 12,
     'quantity_max_length' => 99,
-    'token_expire' => env('TOKEN_EXPIRE', 90),
+    'token_expire' => (int)env('TOKEN_EXPIRE', 90),
 ];
